refactor(plugins): move the plugin entrypoint from `plugins/index.ts` to `plugins.ts`
- Remove `plugins/remark-reading-time.ts`.
- Inline the `remarkReadingTime` implementation.
