+++ plugins.ts
@@ -1,10 +1,11 @@
 import { visit } from 'unist-util-visit'
+import { toString } from 'mdast-util-to-string'
+import getReadingTime from 'reading-time'
 
 import remarkDirective from 'remark-directive'
 import remarkDirectiveSugar from 'remark-directive-sugar'
 import remarkImgattr from 'remark-imgattr'
 import remarkMath from 'remark-math'
-import remarkReadingTime from './remark-reading-time'
 
 import { rehypeHeadingIds } from '@astrojs/markdown-remark'
 import rehypeCallouts from 'rehype-callouts'
@@ -14,11 +15,26 @@ import rehypeAutolinkHeadings from 'rehype-autolink-headings'
 // @ts-expect-error(rehype-wrap-all is not typed)
 import rehypeWrapAll from 'rehype-wrap-all'
 
-import { UI } from '../src/config'
+import { UI } from './src/config'
 
 import type { RemarkPlugins, RehypePlugins } from 'astro'
 import type { PropertiesFromTextDirective } from 'remark-directive-sugar'
 import type { CreateProperties } from 'rehype-external-links'
+
+// https://docs.astro.build/en/recipes/reading-time/
+function remarkReadingTime() {
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-expect-error
+  return (tree, file) => {
+    const { frontmatter } = file.data.astro
+    if (frontmatter.minutesRead || frontmatter.minutesRead === 0) return
+
+    const textOnPage = toString(tree)
+    const readingTime = getReadingTime(textOnPage)
+
+    frontmatter.minutesRead = Math.max(1, Math.round(readingTime.minutes))
+  }
+}
 
 export const remarkPlugins: RemarkPlugins = [
   // https://github.com/remarkjs/remark-directive
